Morph GCE scripts into GCE-coreos scripts
Modifies the forked gce-coreos scripts directory from its origin as
a set of GCE + Debian scripts to a set of GCE + CoreOS scripts.

Since CoreOS doesn't come with Salt and apt-get, this required making
a few changes.  In particular, we've lost a lot of the PKI and
authentication support.  The scripts try to setup an IP-based filter to
offer some protection, but this setup is still very insecure.

See also: the newly added README.md