GCE-CoreOS: Override kubelet's hostname detection
Force the kubelet to listen on an etcd path based on its
non-fully-qualified hostname.

This fixes an issue that prevented the minions from receiving commands
from the master.

Baground: The master's "--cloud_provider=gce" kubelet identification
logic doesn't work on the GCE CoreOS image.  It fails to separate the
specific hostname from the rest of the FQDN.  The result is that the
master publishes events to an etcd path that includes just the machine
name, while the kubelet tries to listen on a path based on the FQDN.

See also: PR #584